remove sumo.views
